+++ lib/super/link_builder.rb
@@ -6,15 +6,6 @@ module Super
       @text = text
       @href = href
       @options = options
-      @requirements = []
-      @requirements += gather_requirements(text)
-      @requirements += gather_requirements(href)
-      @requirements += gather_requirements(options)
-
-      unknown_arguments = @requirements - known_requirements
-      if unknown_arguments.any?
-        raise Error::ArgumentError, "Unknown arguments: #{unknown_arguments.join(", ")}"
-      end
     end
 
     attr_reader :requirements
@@ -28,20 +19,6 @@ module Super
     end
 
     private
-
-    def known_requirements
-      %i[params record].freeze
-    end
-
-    def gather_requirements(value_or_proc)
-      return [] if !value_or_proc.is_a?(Proc)
-
-      requirements =
-        value_or_proc
-          .parameters
-          .select { |(kind, name)| kind = :keyreq }
-          .map(&:last)
-    end
 
     def into_value(value_or_proc, kwargs)
       if value_or_proc.kind_of?(Proc)