+++ lib/super/useful/i19.rb
@@ -1,0 +1,31 @@
+# frozen_string_literal: true
+
+module Super
+  module Useful
+    module I19
+      module_function
+
+      def build_chain(prefix, optional, suffix)
+        prefix = Array(prefix)
+        optional = Array(optional)
+        suffix = Array(suffix)
+
+        result = []
+
+        (optional.size + 1).times do
+          lookup = [prefix, optional, suffix].flatten.join(".").to_sym
+          result.push(lookup)
+          optional.pop
+        end
+
+        result
+      end
+
+      def chain_to_i18n(chain)
+        head, *tail = chain
+
+        [head, { default: tail }]
+      end
+    end
+  end
+end